Feature enhancements (provide more variables to mustache templates)
Provides the following additional variables to mustache templates:

| Variable Name              | Type       | In         | Condition
|
| -------------------------- | ---------- | ---------  |
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
|
| `hasBodyParameters`        | Boolean    | method     | If one or more
parameters are in body (`parameter.in === "body"`)
|
| `hasPathParameters`        | Boolean    | method     | If one or more
parameters are in body (`parameter.in === "path"`)
|
| `hasQueryParameters`       | Boolean    | method     | If one or more
parameters are in body (`parameter.in === "query"`)
|
| `hasHeaderParameters`      | Boolean    | method     | If one or more
parameters are in body (`parameter.in === "header"`)
|
| `hasFormDataParameters`    | Boolean    | method     | If one or more
parameters are in body (`parameter.in === "formData"`)
|
| `hasFormDataParameters`    | Boolean    | method     | If one or more
parameters are in body (`parameter.in === "formData"`)
|
| `hasBodyParameters`        | Boolean    | parameter  | If a schema is
present and is of type "object" (`parameter.schema.type === "object"`)
|
| `bodyParams`               | Array      | parameter  | If
`parameter.hasBodyParameters === true` and parameter schema properties
are defined. Array contains objects are in [JSON Schema
v4.0](http://json-schema.org/) format with an added parameter `name`. |
| `hasSingularBodyParameter` | Boolean    | parameter  | If a schema is
present and is of type "number", "string", or "array"
(`parameter.schema.type === "number/string/array"`)
|
| `bodyParam`                | Object     | parameter  | If
`parameter.hasSingularBodyParameter === true`. Only contains the type of
the parameter.
|
| `typeIn`                   | Array      | bodyParam  | If `bodyParams`
are included and the defined type is an object or array of types, rather
than a single type.
|
| `responses`                | Array      | method     | If responses
are defined for the method.
|
| `security`                 | Array      | method     | If security
definitions are included for the method.
|